add : 240619_future 함수 과제
