--- lib/13_functional programming/training.dart
@@ -1,29 +1,0 @@
-class Trader {
-  String name;
-  String city;
-
-  Trader(this.name, this.city);
-}
-
-class Transaction {
-  Trader trader;
-  int year;
-  int value;
-
-  Transaction(this.trader, this.year, this.value);
-}
-
-final transactions = [
-  Transaction(Trader("Brian", "Cambridge"), 2011, 300),
-  Transaction(Trader("Raoul", "Cambridge"), 2012, 1000),
-  Transaction(Trader("Raoul", "Cambridge"), 2011, 400),
-  Transaction(Trader("Mario", "Milan"), 2012, 710),
-  Transaction(Trader("Mario", "Milan"), 2012, 700),
-  Transaction(Trader("Alan", "Cambridge"), 2012, 950),
-];
-
-//정규표현식
-bool isValidPlayerName(String name) {
-  return RegExp(r'[A-Z][A-Z0-9]{7}').hasMatch(name);
-}
-r : row
