+++ lib/14_async/sample.csv
@@ -1,0 +1,2 @@
+홍길동, 30
+한석봉, 20
